Remove CAPI
