+++ .github/workflows/build-package.yml
@@ -10,16 +10,14 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - name: Checkout 🛎️
-      uses: actions/checkout@v3
+      uses: actions/checkout@v4
       with:
         persist-credentials: false
-    - uses: actions/setup-dotnet@v3
+    - uses: actions/setup-dotnet@v4
       with:
         dotnet-version: '6.x'
     - name: Build the solution 🔨
       run: dotnet build
-    - name: Run Unit Tests 🍴
-      run: dotnet test
     - name: Create NuGet package 🌎
       run: dotnet pack DolbyIO.Rest/DolbyIO.Rest.csproj -p:Configuration=Release
     - name: Sign NuGet Package
@@ -29,7 +27,7 @@ jobs:
         dotnet nuget sign DolbyIO.Rest.*.nupkg --certificate-path ./certificate.pfx --certificate-password ${{ secrets.WINDOWS_CERTIFICATE_PASSWORD}} --timestamper http://timestamp.digicert.com/
         rm certificate.pfx
     - name: Upload artifacts
-      uses: actions/upload-artifact@v3
+      uses: actions/upload-artifact@v4
       with:
         name: nuget-package
         path: DolbyIO.Rest/bin/Release
