--- DolbyIO.Rest/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleToAttribute("DolbyIO.Rest.Tests")]
