--- DolbyIO.Rest/Communications/Models/Recording.cs
@@ -1,9 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace DolbyIO.Rest.Communications.Models;
-
-internal sealed class StartRecordingRequest
-{
-    [JsonProperty("layoutUrl")]
-    public string LayoutUrl { get; set; }
-}
