+++ DolbyIO.Rest/Media/Jobs.cs
@@ -4,7 +4,6 @@ using System.Collections.Specialized;
 using System.Net.Http;
 using System.Threading.Tasks;
 using DolbyIO.Rest.Media.Models;
-using DolbyIO.Rest.Models;
 
 namespace DolbyIO.Rest.Media;
