+++ DolbyIO.Rest/Media/Models/JwtToken.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace DolbyIO.Rest.Models;
+namespace DolbyIO.Rest.Media.Models;
 
 public sealed class JwtToken
 {
@@ -21,12 +21,4 @@ public sealed class JwtToken
     /// </summary>
     [JsonProperty("token_type")]
     public string TokenType { get; internal set; }
-
-#nullable enable
-    /// <summary>
-    /// Gets the scope of access token.
-    /// </summary>
-    [JsonProperty("scope")]
-    public string? Scope { get; internal set; }
-#nullable disable
 }
