+++ DolbyIO.Rest/Urls.cs
@@ -3,9 +3,7 @@ namespace DolbyIO.Rest;
 internal static class Urls
 {
     public const string API_BASE_URL = "https://api.dolby.io";
-    public const string SESSION_BASE_URL = "https://session.voxeet.com";
-    public const string CAPI_BASE_URL = "https://comms.api.dolby.io";
     public const string MAPI_BASE_URL = "https://api.dolby.com";
-    public const string SAPI_BASE_URL = "https://api.millicast.com";
-    public const string SAPI_DIRECTOR_BASE_URL = "https://director.millicast.com";
+    public const string STREAMING_BASE_URL = "https://api.millicast.com";
+    public const string STREAMING_DIRECTOR_BASE_URL = "https://director.millicast.com";
 }
